Divide examples to speed up ci
