+++ .github/workflows/e2e-kubernetes.yaml
@@ -6,7 +6,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        TEST_GROUP: [smoke, es, cassandra, streaming, examples]
+        TEST_GROUP: [smoke, es, cassandra, streaming, examples1, examples2]
     steps:
     - uses: jpkrohling/setup-minikube@v1-release
     - uses: jpkrohling/setup-kubectl@v1-release
