+++ test/e2e/utils.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"net/http"
 	"os"
+	"os/exec"
 	"runtime"
 	"strings"
 	"testing"
@@ -262,4 +263,39 @@ type services struct {
 	limit  int         `json:"limit"`
 	offset int         `json:offset`
 	errors interface{} `json:"errors"`
+}
+
+func createJaegerInstanceFromFile(name, filename string) *v1.Jaeger {
+	cmd := exec.Command("kubectl", "create", "--namespace", namespace, "--filename", filename)
+	output, err := cmd.CombinedOutput()
+	if err != nil && !strings.Contains(string(output), "AlreadyExists") {
+		require.NoError(t, err, "Failed creating Jaeger instance with: [%s]\n", string(output))
+	}
+
+	return getJaegerInstance(name, namespace)
+}
+
+func smokeTestAllInOneExample(name, yamlFileName string) {
+	jaegerInstance := createJaegerInstanceFromFile(name, yamlFileName)
+	defer undeployJaegerInstance(jaegerInstance)
+
+	err := e2eutil.WaitForDeployment(t, fw.KubeClient, namespace, name, 1, retryInterval, timeout)
+	require.NoErrorf(t, err, "Error waiting for %s to deploy", name)
+
+	AllInOneSmokeTest(name)
+}
+
+func smokeTestProductionExample(name, yamlFileName string) {
+	jaegerInstance := createJaegerInstanceFromFile(name, yamlFileName)
+	defer undeployJaegerInstance(jaegerInstance)
+
+	queryDeploymentName := name + "-query"
+	collectorDeploymentName := name + "-collector"
+
+	err := e2eutil.WaitForDeployment(t, fw.KubeClient, namespace, queryDeploymentName, 1, retryInterval, timeout)
+	require.NoErrorf(t, err, "Error waiting for %s to deploy", queryDeploymentName)
+	err = e2eutil.WaitForDeployment(t, fw.KubeClient, namespace, collectorDeploymentName, 1, retryInterval, timeout)
+	require.NoErrorf(t, err, "Error waiting for %s to deploy", collectorDeploymentName)
+
+	ProductionSmokeTest(name)
 }
